[Refactor] Oauth Refactor#73

+++ src/main/java/com/midnear/midnearshopping/mapper/coupon_point/PointMapper.java
@@ -2,7 +2,6 @@ package com.midnear.midnearshopping.mapper.coupon_point;
 
 import com.midnear.midnearshopping.domain.dto.coupon_point.UserPointDto;
 import com.midnear.midnearshopping.domain.vo.coupon_point.PointVo;
-import com.midnear.midnearshopping.domain.vo.coupon_point.ReviewPointVo;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
@@ -10,6 +9,6 @@ import java.util.List;
 @Mapper
 public interface PointMapper {
     void grantPoints(PointVo pointVo);
-    List<UserPointDto> getPointList(Integer userId);
+    List<UserPointDto> getPointList(Long userId);
     String getProductInfoByReviewId(Long reviewId);
 }
